+++ LABA4/LABA4/main.cpp
@@ -7,12 +7,12 @@ int main() {
     C1.stepen();
     comel<double> C2(1.5, 3);
     C2.stepen();
-    matric <int> M1(2);
+    matric <double> M1(2);
     M1.cinmatr();
     cout << "Âòîðîé ìàññèâ" << endl;
-    matric <int> M2(2);
+    matric <double> M2(2);
     M2.cinmatr();
-     matric <int> M3 = M1.proizved(M2);
+     matric <double> M3 = M1.proizved(M2);
      cout << "ÂÛÂÎÄ ÏÐÎÈÇÂÅÄÅÍÈß: " << endl;
      M3.print();
